got ride of connectors bindings 1
