--- ConnectorRevit/ConnectorRevit/UI/ConnectorBindingsRevit.ClientOperations.cs
@@ -1,120 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Autodesk.Revit.DB;
-using Speckle.ConnectorRevit.Storage;
-using Speckle.Core.Api;
-using Speckle.Core.Kits;
-using Speckle.Core.Logging;
-using Speckle.Core.Models;
-using Speckle.Core.Transports;
-using Speckle.DesktopUI.Utils;
-using Stylet;
-using RevitElement = Autodesk.Revit.DB.Element;
-
-namespace Speckle.ConnectorRevit.UI
-{
-  public partial class ConnectorBindingsRevit
-  {
-
-
-    public List<StreamState> DocumentStreams { get; set; } = new List<StreamState>();
-
-    [ObsoleteAttribute("This property is obsolete. Use the Report on the progress view model", false)]
-    public List<Exception> ConversionErrors { get; set; } = new List<Exception>();
-
-    /// <summary>
-    /// Keeps track of errors in the operations of send/receive.
-    /// </summary>
-    [ObsoleteAttribute("This property is obsolete. Use the Report on the progress view model", false)]
-    public List<Exception> OperationErrors { get; set; } = new List<Exception>();
-
-    public override List<StreamState> GetStreamsInFile()
-    {
-      if (CurrentDoc != null)
-        DocumentStreams = StreamStateManager.ReadState(CurrentDoc.Document);
-
-      return DocumentStreams;
-    }
-
-    #region Local file i/o
-
-    /// <summary>
-    /// Adds a new stream to the file.
-    /// </summary>
-    /// <param name="state">StreamState passed by the UI</param>
-    public override void AddNewStream(StreamState state)
-    {
-      var index = DocumentStreams.FindIndex(b => b.Stream.id == state.Stream.id);
-      if (index == -1)
-      {
-        DocumentStreams.Add(state);
-        WriteStateToFile();
-      }
-    }
-
-    /// <summary>
-    /// Removes a stream from the file.
-    /// </summary>
-    /// <param name="streamId"></param>
-    public override void RemoveStreamFromFile(string streamId)
-    {
-      var streamState = DocumentStreams.FirstOrDefault(s => s.Stream.id == streamId);
-      if (streamState != null)
-      {
-        DocumentStreams.Remove(streamState);
-        WriteStateToFile();
-      }
-    }
-
-    /// <summary>
-    /// Update the stream state and adds adds the filtered objects
-    /// </summary>
-    /// <param name="state"></param>
-    public override void PersistAndUpdateStreamInFile(StreamState state)
-    {
-      var index = DocumentStreams.FindIndex(b => b.Stream.id == state.Stream.id);
-      if (index != -1)
-      {
-        DocumentStreams[index] = state;
-        WriteStateToFile();
-      }
-    }
-
-    /// <summary>
-    /// Transaction wrapper around writing the local streams to the file.
-    /// </summary>
-    private void WriteStateToFile()
-    {
-      Queue.Add(new Action(() =>
-      {
-        using (Transaction t = new Transaction(CurrentDoc.Document, "Speckle Write State"))
-        {
-          t.Start();
-          StreamStateManager.WriteStreamStateList(CurrentDoc.Document, DocumentStreams);
-          t.Commit();
-        }
-      }));
-      Executor.Raise();
-    }
-
-    #endregion
-
-    private void UpdateProgress(ConcurrentDictionary<string, int> dict, DesktopUI.Utils.ProgressReport progress)
-    {
-      if (progress == null)
-      {
-        return;
-      }
-
-      Execute.PostToUIThread(() =>
-      {
-        progress.ProgressDict = dict;
-        progress.Value = dict.Values.Last();
-      });
-    }
-  }
-}
